refactor: move NDVI cache to cache directory
