+++ backend/src/cache/ndvi_cache.py
